+++ Foundation/GTMNSData+zlibTest.m
@@ -105,12 +105,12 @@ static BOOL HasGzipHeader(NSData *data) {
   STAssertNil([NSData gtm_dataByInflatingData:data], nil);
   
   // test deflated data runs that end before they are done
-  for (int x = 1 ; x < [deflated length] ; ++x) {
+  for (NSUInteger x = 1 ; x < [deflated length] ; ++x) {
     STAssertNil([NSData gtm_dataByInflatingBytes:[deflated bytes] length:x], nil);
   }
   
   // test gzipped data runs that end before they are done
-  for (int x = 1 ; x < [gzipped length] ; ++x) {
+  for (NSUInteger x = 1 ; x < [gzipped length] ; ++x) {
     STAssertNil([NSData gtm_dataByInflatingBytes:[gzipped bytes] length:x], nil);
   }
   
@@ -175,21 +175,21 @@ static BOOL HasGzipHeader(NSData *data) {
       // w/ *Bytes apis, default level
       NSData *deflated = [NSData gtm_dataByDeflatingBytes:[data bytes] length:[data length]];
       STAssertNotNil(deflated, @"failed to deflate data block");
-      STAssertGreaterThan([deflated length], 0U, @"failed to deflate data block");
+      STAssertGreaterThan([deflated length], (NSUInteger)0, @"failed to deflate data block");
       STAssertFalse(HasGzipHeader(deflated), @"has gzip header on zlib data");
       NSData *dataPrime = [NSData gtm_dataByInflatingBytes:[deflated bytes] length:[deflated length]];
       STAssertNotNil(dataPrime, @"failed to inflate data block");
-      STAssertGreaterThan([dataPrime length], 0U, @"failed to inflate data block");
+      STAssertGreaterThan([dataPrime length], (NSUInteger)0, @"failed to inflate data block");
       STAssertEqualObjects(data, dataPrime, @"failed to round trip via *Bytes apis");
 
       // w/ *Data apis, default level
       deflated = [NSData gtm_dataByDeflatingData:data];
       STAssertNotNil(deflated, @"failed to deflate data block");
-      STAssertGreaterThan([deflated length], 0U, @"failed to deflate data block");
+      STAssertGreaterThan([deflated length], (NSUInteger)0, @"failed to deflate data block");
       STAssertFalse(HasGzipHeader(deflated), @"has gzip header on zlib data");
       dataPrime = [NSData gtm_dataByInflatingData:deflated];
       STAssertNotNil(dataPrime, @"failed to inflate data block");
-      STAssertGreaterThan([dataPrime length], 0U, @"failed to inflate data block");
+      STAssertGreaterThan([dataPrime length], (NSUInteger)0, @"failed to inflate data block");
       STAssertEqualObjects(data, dataPrime, @"failed to round trip via *Data apis");
 
       // loop over the compression levels
@@ -199,21 +199,21 @@ static BOOL HasGzipHeader(NSData *data) {
                                              length:[data length]
                                    compressionLevel:level];
         STAssertNotNil(deflated, @"failed to deflate data block");
-        STAssertGreaterThan([deflated length], 0U, @"failed to deflate data block");
+        STAssertGreaterThan([deflated length], (NSUInteger)0, @"failed to deflate data block");
         STAssertFalse(HasGzipHeader(deflated), @"has gzip header on zlib data");
         dataPrime = [NSData gtm_dataByInflatingBytes:[deflated bytes] length:[deflated length]];
         STAssertNotNil(dataPrime, @"failed to inflate data block");
-        STAssertGreaterThan([dataPrime length], 0U, @"failed to inflate data block");
+        STAssertGreaterThan([dataPrime length], (NSUInteger)0, @"failed to inflate data block");
         STAssertEqualObjects(data, dataPrime, @"failed to round trip via *Bytes apis");
 
         // w/ *Data apis, using our level
         deflated = [NSData gtm_dataByDeflatingData:data compressionLevel:level];
         STAssertNotNil(deflated, @"failed to deflate data block");
-        STAssertGreaterThan([deflated length], 0U, @"failed to deflate data block");
+        STAssertGreaterThan([deflated length], (NSUInteger)0, @"failed to deflate data block");
         STAssertFalse(HasGzipHeader(deflated), @"has gzip header on zlib data");
         dataPrime = [NSData gtm_dataByInflatingData:deflated];
         STAssertNotNil(dataPrime, @"failed to inflate data block");
-        STAssertGreaterThan([dataPrime length], 0U, @"failed to inflate data block");
+        STAssertGreaterThan([dataPrime length], (NSUInteger)0, @"failed to inflate data block");
         STAssertEqualObjects(data, dataPrime, @"failed to round trip via *Data apis");
       }
 
@@ -240,21 +240,21 @@ static BOOL HasGzipHeader(NSData *data) {
       // w/ *Bytes apis, default level
       NSData *gzipped = [NSData gtm_dataByGzippingBytes:[data bytes] length:[data length]];
       STAssertNotNil(gzipped, @"failed to gzip data block");
-      STAssertGreaterThan([gzipped length], 0U, @"failed to gzip data block");
+      STAssertGreaterThan([gzipped length], (NSUInteger)0, @"failed to gzip data block");
       STAssertTrue(HasGzipHeader(gzipped), @"doesn't have gzip header on gzipped data");
       NSData *dataPrime = [NSData gtm_dataByInflatingBytes:[gzipped bytes] length:[gzipped length]];
       STAssertNotNil(dataPrime, @"failed to inflate data block");
-      STAssertGreaterThan([dataPrime length], 0U, @"failed to inflate data block");
+      STAssertGreaterThan([dataPrime length], (NSUInteger)0, @"failed to inflate data block");
       STAssertEqualObjects(data, dataPrime, @"failed to round trip via *Bytes apis");
 
       // w/ *Data apis, default level
       gzipped = [NSData gtm_dataByGzippingData:data];
       STAssertNotNil(gzipped, @"failed to gzip data block");
-      STAssertGreaterThan([gzipped length], 0U, @"failed to gzip data block");
+      STAssertGreaterThan([gzipped length], (NSUInteger)0, @"failed to gzip data block");
       STAssertTrue(HasGzipHeader(gzipped), @"doesn't have gzip header on gzipped data");
       dataPrime = [NSData gtm_dataByInflatingData:gzipped];
       STAssertNotNil(dataPrime, @"failed to inflate data block");
-      STAssertGreaterThan([dataPrime length], 0U, @"failed to inflate data block");
+      STAssertGreaterThan([dataPrime length], (NSUInteger)0, @"failed to inflate data block");
       STAssertEqualObjects(data, dataPrime, @"failed to round trip via *Data apis");
 
       // loop over the compression levels
@@ -264,21 +264,21 @@ static BOOL HasGzipHeader(NSData *data) {
                                            length:[data length]
                                  compressionLevel:level];
         STAssertNotNil(gzipped, @"failed to gzip data block");
-        STAssertGreaterThan([gzipped length], 0U, @"failed to gzip data block");
+        STAssertGreaterThan([gzipped length], (NSUInteger)0, @"failed to gzip data block");
         STAssertTrue(HasGzipHeader(gzipped), @"doesn't have gzip header on gzipped data");
         dataPrime = [NSData gtm_dataByInflatingBytes:[gzipped bytes] length:[gzipped length]];
         STAssertNotNil(dataPrime, @"failed to inflate data block");
-        STAssertGreaterThan([dataPrime length], 0U, @"failed to inflate data block");
+        STAssertGreaterThan([dataPrime length], (NSUInteger)0, @"failed to inflate data block");
         STAssertEqualObjects(data, dataPrime, @"failed to round trip via *Bytes apis");
 
         // w/ *Data apis, using our level
         gzipped = [NSData gtm_dataByGzippingData:data compressionLevel:level];
         STAssertNotNil(gzipped, @"failed to gzip data block");
-        STAssertGreaterThan([gzipped length], 0U, @"failed to gzip data block");
+        STAssertGreaterThan([gzipped length], (NSUInteger)0, @"failed to gzip data block");
         STAssertTrue(HasGzipHeader(gzipped), @"doesn't have gzip header on gzipped data");
         dataPrime = [NSData gtm_dataByInflatingData:gzipped];
         STAssertNotNil(dataPrime, @"failed to inflate data block");
-        STAssertGreaterThan([dataPrime length], 0U, @"failed to inflate data block");
+        STAssertGreaterThan([dataPrime length], (NSUInteger)0, @"failed to inflate data block");
         STAssertEqualObjects(data, dataPrime, @"failed to round trip via *Data apis");
       }
